+++ jupiter-common/src/main/java/org/jupiter/common/util/ExceptionUtil.java
@@ -16,7 +16,7 @@
 
 package org.jupiter.common.util;
 
-import org.jupiter.common.util.internal.JUnsafe;
+import org.jupiter.common.util.internal.UnsafeUtil;
 import org.jupiter.common.util.internal.UnsafeReferenceFieldUpdater;
 import org.jupiter.common.util.internal.UnsafeUpdater;
 import sun.misc.Unsafe;
@@ -36,7 +36,7 @@ public class ExceptionUtil {
      * Raises an exception bypassing compiler checks for checked exceptions.
      */
     public static void throwException(Throwable t) {
-        Unsafe unsafe = JUnsafe.getUnsafe();
+        Unsafe unsafe = UnsafeUtil.getUnsafe();
         if (unsafe != null) {
             unsafe.throwException(t);
         } else {