+++ jupiter-common/src/test/java/org/jupiter/common/util/GetFieldValueBenchmark.java
@@ -16,7 +16,7 @@
 
 package org.jupiter.common.util;
 
-import org.jupiter.common.util.internal.JUnsafe;
+import org.jupiter.common.util.internal.UnsafeUtil;
 import org.openjdk.jmh.annotations.*;
 import org.openjdk.jmh.runner.Runner;
 import org.openjdk.jmh.runner.RunnerException;
@@ -59,7 +59,7 @@ public class GetFieldValueBenchmark {
     @Benchmark
     public void unsafeGet() {
         // RandomLoadBalancer中有类似代码
-        Object[] array = (Object[]) JUnsafe.getUnsafe().getObjectVolatile(FIELD_TEST, FieldTest.OFFSET);
+        Object[] array = (Object[]) UnsafeUtil.getUnsafe().getObjectVolatile(FIELD_TEST, FieldTest.OFFSET);
         if ((int) array[0] != 1) {
             System.out.println(1);
         }
@@ -78,7 +78,7 @@ class FieldTest {
     static {
         long offsetTmp;
         try {
-            offsetTmp = JUnsafe.getUnsafe().objectFieldOffset(Reflects.getField(FieldTest.class, "array"));
+            offsetTmp = UnsafeUtil.getUnsafe().objectFieldOffset(Reflects.getField(FieldTest.class, "array"));
         } catch (NoSuchFieldException e) {
             offsetTmp = 0;
         }