+++ jupiter-serialization/jupiter-serialization-protostuff/src/main/java/org/jupiter/serialization/proto/buffer/UnsafeNioBufOutput.java
@@ -16,12 +16,15 @@
 
 package org.jupiter.serialization.proto.buffer;
 
+import org.jupiter.common.util.internal.UnsafeDirectBufferUtil;
+import org.jupiter.common.util.internal.UnsafeUtf8Util;
+import org.jupiter.common.util.internal.UnsafeUtil;
 import org.jupiter.serialization.OutputBuf;
-import sun.nio.ch.DirectBuffer;
 
 import java.io.IOException;
 
-import static org.jupiter.common.util.internal.UnsafeByteBufferUtil.*;
+import static io.protostuff.WireFormat.WIRETYPE_LENGTH_DELIMITED;
+import static io.protostuff.WireFormat.makeTag;
 
 /**
  * jupiter
@@ -43,6 +46,39 @@ class UnsafeNioBufOutput extends NioBufOutput {
     }
 
     @Override
+    public void writeString(int fieldNumber, CharSequence value, boolean repeated) throws IOException {
+        writeVarInt32(makeTag(fieldNumber, WIRETYPE_LENGTH_DELIMITED));
+
+        int position = nioBuffer.position();
+        // UTF-8 byte length of the string is at least its UTF-16 code unit length (value.length()),
+        // and at most 3 times of it. We take advantage of this in both branches below.
+        int maxLengthVarIntSize = computeVarInt32Size(value.length() * UnsafeUtf8Util.MAX_BYTES_PER_CHAR);
+        int minLengthVarIntSize = computeVarInt32Size(value.length());
+        if (maxLengthVarIntSize == minLengthVarIntSize) {
+            // Save the current position and increment past the length field. We'll come back
+            // and write the length field after the encoding is complete.
+            int stringStartPos = position + minLengthVarIntSize;
+            nioBuffer.position(stringStartPos);
+
+            // Encode the string.
+            UnsafeUtf8Util.encodeUtf8Direct(value, nioBuffer);
+
+            // Write the length and advance the position.
+            int length = nioBuffer.position() - stringStartPos;
+            nioBuffer.position(position);
+            writeVarInt32(length);
+            nioBuffer.position(stringStartPos + length);
+        } else {
+            // Calculate and write the encoded length.
+            int length = UnsafeUtf8Util.encodedLength(value);
+            writeVarInt32(length);
+
+            // Write the string and advance the position.
+            UnsafeUtf8Util.encodeUtf8Direct(value, nioBuffer);
+        }
+    }
+
+    @Override
     protected void writeVarInt32(int value) throws IOException {
         byte[] buf = new byte[5];
         int locPtr = 0;
@@ -51,7 +87,7 @@ class UnsafeNioBufOutput extends NioBufOutput {
             if ((value & ~0x7F) == 0) {
                 buf[locPtr++] = (byte) value;
                 ensureCapacity(locPtr);
-                setBytes(address(position), buf, 0, locPtr);
+                UnsafeDirectBufferUtil.setBytes(address(position), buf, 0, locPtr);
                 nioBuffer.position(position + locPtr);
                 return;
             } else {
@@ -70,7 +106,7 @@ class UnsafeNioBufOutput extends NioBufOutput {
             if ((value & ~0x7FL) == 0) {
                 buf[locPtr++] = (byte) value;
                 ensureCapacity(locPtr);
-                setBytes(address(position), buf, 0, locPtr);
+                UnsafeDirectBufferUtil.setBytes(address(position), buf, 0, locPtr);
                 nioBuffer.position(position + locPtr);
                 return;
             } else {
@@ -84,7 +120,7 @@ class UnsafeNioBufOutput extends NioBufOutput {
     protected void writeInt32LE(int value) throws IOException {
         ensureCapacity(4);
         int position = nioBuffer.position();
-        setIntLE(address(position), value);
+        UnsafeDirectBufferUtil.setIntLE(address(position), value);
         nioBuffer.position(position + 4);
     }
 
@@ -92,7 +128,7 @@ class UnsafeNioBufOutput extends NioBufOutput {
     protected void writeInt64LE(long value) throws IOException {
         ensureCapacity(8);
         int position = nioBuffer.position();
-        setLongLE(address(position), value);
+        UnsafeDirectBufferUtil.setLongLE(address(position), value);
         nioBuffer.position(position + 8);
     }
 
@@ -100,7 +136,7 @@ class UnsafeNioBufOutput extends NioBufOutput {
     protected void writeByte(byte value) throws IOException {
         ensureCapacity(1);
         int position = nioBuffer.position();
-        setByte(address(position), value);
+        UnsafeDirectBufferUtil.setByte(address(position), value);
         nioBuffer.position(position + 1);
     }
 
@@ -108,7 +144,7 @@ class UnsafeNioBufOutput extends NioBufOutput {
     protected void writeByteArray(byte[] value, int offset, int length) throws IOException {
         ensureCapacity(length);
         int position = nioBuffer.position();
-        setBytes(address(position), value, offset, length);
+        UnsafeDirectBufferUtil.setBytes(address(position), value, offset, length);
         nioBuffer.position(position + length);
     }
 
@@ -132,10 +168,29 @@ class UnsafeNioBufOutput extends NioBufOutput {
     }
 
     private void updateBufferAddress() {
-        memoryAddress = ((DirectBuffer) nioBuffer).address();
+        memoryAddress = UnsafeUtil.addressOffset(nioBuffer);
     }
 
     private long address(int position) {
         return memoryAddress + position;
+    }
+
+    /**
+     * Compute the number of bytes that would be needed to encode a {@code uint32} field.
+     */
+    private static int computeVarInt32Size(final int value) {
+        if ((value & (~0 <<  7)) == 0) {
+            return 1;
+        }
+        if ((value & (~0 << 14)) == 0) {
+            return 2;
+        }
+        if ((value & (~0 << 21)) == 0) {
+            return 3;
+        }
+        if ((value & (~0 << 28)) == 0) {
+            return 4;
+        }
+        return 5;
     }
 }